% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifyop.R
\name{\%<->\%}
\alias{\%<->\%}
\title{unify operator}
\usage{
\%>\%(lhs, rhs) <- value
}
\arguments{
\item{lhs}{A vector of variables to be bound to corresponding values in right hand side}

\item{rhs}{A vector or list of values to be copied to the left hand side}
}
\description{
The unify operator takes on its left hand side a vector of variable names to be bound
to the corresponding values on the right hand side.
}
\examples{
tmp1 <- c(10, 11, 12)
c(h, i, j) \%<->\% tmp1

cat1 <- list(d1 = "hej", d2 = "hopp", d3 = "lingon")
c(k, l, m) \%<->\% cat1

c(a, b, c) \%<->\% c(10, 11, 12)
}

